test: Refactor time tests
- Move testCurrentDateTimeFunction from JodaTimeDefaultsTest to JodaTimeTests
- Rewrite testDefaultCurrentDateTime in JodaTimeDefaultsTest
- Add testDefaultCurrentDateTime to Java and Kotlin tests
