+++ exposed-tests/src/main/kotlin/org/jetbrains/exposed/sql/tests/TestUtils.kt
@@ -1,6 +1,8 @@
 package org.jetbrains.exposed.sql.tests
 
 import org.jetbrains.exposed.sql.Column
+import org.jetbrains.exposed.sql.Table
+import org.jetbrains.exposed.sql.insert
 import org.jetbrains.exposed.sql.transactions.TransactionManager
 import org.jetbrains.exposed.sql.vendors.DatabaseDialect
 import org.jetbrains.exposed.sql.vendors.SQLServerDialect
@@ -10,10 +12,11 @@ fun String.inProperCase(): String = TransactionManager.currentOrNull()?.db?.iden
 
 val currentDialectTest: DatabaseDialect get() = TransactionManager.current().db.dialect
 
-val currentDialectIfAvailableTest: DatabaseDialect? get() =
-    if (TransactionManager.isInitialized() && TransactionManager.currentOrNull() != null) {
-        currentDialectTest
-    } else null
+val currentDialectIfAvailableTest: DatabaseDialect?
+    get() =
+        if (TransactionManager.isInitialized() && TransactionManager.currentOrNull() != null) {
+            currentDialectTest
+        } else null
 
 inline fun <reified E : Enum<E>> enumSetOf(vararg elements: E): EnumSet<E> =
     elements.toCollection(EnumSet.noneOf(E::class.java))
@@ -21,3 +24,9 @@ inline fun <reified E : Enum<E>> enumSetOf(vararg elements: E): EnumSet<E> =
 fun <T> Column<T>.constraintNamePart() = (currentDialectTest as? SQLServerDialect)?.let {
     " CONSTRAINT DF_${table.tableName}_$name"
 } ?: ""
+
+fun Table.insertAndWait(duration: Long) {
+    this.insert { }
+    TransactionManager.current().commit()
+    Thread.sleep(duration)
+}
